fix: update type of ListQuery.Limit to int32
To match the type (int32) of the AIP page size standard field.

+++ internal/codegen/databasecodegen/testdata/1.sql.database.go
@@ -93,7 +93,7 @@ func (t SingersReadTransaction) List(
 			Offset: spansql.Param("offset"),
 		}.SQL(),
 		Params: map[string]interface{}{
-			"limit":  query.Limit,
+			"limit":  int64(query.Limit),
 			"offset": query.Offset,
 		},
 	}
@@ -304,7 +304,7 @@ func (r *SingersRow) Key() SingersKey {
 type ListQuery struct {
 	Where  spansql.BoolExpr
 	Order  []spansql.Order
-	Limit  int64
+	Limit  int32
 	Offset int64
 }
