+++ internal/examples/musicdb/database_gen.go
@@ -95,7 +95,7 @@ func (t SingersReadTransaction) List(
 			Offset: spansql.Param("offset"),
 		}.SQL(),
 		Params: map[string]interface{}{
-			"limit":  query.Limit,
+			"limit":  int64(query.Limit),
 			"offset": query.Offset,
 		},
 	}
@@ -360,7 +360,7 @@ func (t SingersAndAlbumsReadTransaction) List(
 	stmt := spanner.Statement{
 		SQL: q.String(),
 		Params: map[string]interface{}{
-			"limit":  query.Limit,
+			"limit":  int64(query.Limit),
 			"offset": query.Offset,
 		},
 	}
@@ -406,7 +406,7 @@ func (t SingersAndAlbumsReadTransaction) BatchGet(
 	foundRows := make(map[SingersKey]*SingersAndAlbumsRow, len(keys))
 	if err := t.List(ctx, ListQuery{
 		Where: spansql.Paren{Expr: where},
-		Limit: int64(len(keys)),
+		Limit: int32(len(keys)),
 	}).Do(func(row *SingersAndAlbumsRow) error {
 		foundRows[row.Key()] = row
 		return nil
@@ -599,7 +599,7 @@ func (t AlbumsReadTransaction) List(
 			Offset: spansql.Param("offset"),
 		}.SQL(),
 		Params: map[string]interface{}{
-			"limit":  query.Limit,
+			"limit":  int64(query.Limit),
 			"offset": query.Offset,
 		},
 	}
@@ -873,7 +873,7 @@ func (t AlbumsAndSongsReadTransaction) List(
 	stmt := spanner.Statement{
 		SQL: q.String(),
 		Params: map[string]interface{}{
-			"limit":  query.Limit,
+			"limit":  int64(query.Limit),
 			"offset": query.Offset,
 		},
 	}
@@ -919,7 +919,7 @@ func (t AlbumsAndSongsReadTransaction) BatchGet(
 	foundRows := make(map[AlbumsKey]*AlbumsAndSongsRow, len(keys))
 	if err := t.List(ctx, ListQuery{
 		Where: spansql.Paren{Expr: where},
-		Limit: int64(len(keys)),
+		Limit: int32(len(keys)),
 	}).Do(func(row *AlbumsAndSongsRow) error {
 		foundRows[row.Key()] = row
 		return nil
@@ -1107,7 +1107,7 @@ func (t SongsReadTransaction) List(
 			Offset: spansql.Param("offset"),
 		}.SQL(),
 		Params: map[string]interface{}{
-			"limit":  query.Limit,
+			"limit":  int64(query.Limit),
 			"offset": query.Offset,
 		},
 	}
@@ -1356,7 +1356,7 @@ func (r *SongsRow) Key() SongsKey {
 type ListQuery struct {
 	Where  spansql.BoolExpr
 	Order  []spansql.Order
-	Limit  int64
+	Limit  int32
 	Offset int64
 }
